+++ modules/tinymce/CHANGELOG.md
@@ -60,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `formatselect` toolbar button and `blockformats` menu item to `blocks` #TINY-8328
 - Renamed `fontsize_formats` option to `font_size_formats`, and `fontsizeselect` toolbar button and `fontsizes` menu item to `fontsize` #TINY-8328
 - Renamed `lineheight_formats` option to `line_height_formats` #TINY-8328
-
+- Renamed `styleselect` toolbar button and `formats` menu item to `styles` #TINY-8328
 
 ### Fixed
 - The object returned from the `editor.fire()` API was incorrect if the editor had been removed #TINY-8018